Swap order of fuses:update/1 mapping
fuses:update/1 now expects a list of {Fuse :: integer(), Name :: term()}
and validates this requirement.

The FuseToName mapping is updated from this list. The NameToFuse mapping
is now built as the invers of FuseToName rather than updating with the
supplied update. This allows fuses to be renamed with out leaving the
old name in the NameToFuse map.