Gain access to flip-flop outputs
For experiment:compile/1 to route a flip-flop's output to another's input
that signal needs to be specifically defined. An output signal is
already defined when an OE buffer is generated, so use that side affect
by compiling {net, _, _, #{oe => true}} to produce the net_Q signal
without using the OE feature of the macro-cell.