Ensure inputs are always used to remove redudant diffs
Funtion block input selection fuses where clouding the diff-matrix,
make sure the contol flip-flop uses all of the inputs in the experiment
to remove those differences.
Add a "d SET s" experiment to highlight the s-term.
Use the new fuse_map to show a name for all fuses in the diff-matrix.