The pt-enable fuses where std-power fuses!!!
Rename
 - power_a to std_power,
 - power_b to pt3_std_power,
 - pt#_enable fuses to pt#_std_power.
The pt#_std_power fuses are only applicable when the product term is in
use, hence the initial lablling as "enable" fuses.
The std_power fuse appears to be for the logic & flip-flop components
and not needed when those components are not in use, i.e. when product
terms are being forwarded.