Clarify clock selection fuses
Rename gck_1 & gck_0 to clk_mux0 & clk_mux1 since they are not specific
to the global clocks. Update old experiment & databases to match.

Order fuses was changed to match they occur in the fuse-map, mux0 for
the first and mux1 for the second. This was also applied to other mux
pairs currently marked as a "guess".

Add clk_confirmation module that confirms these fuses and the pt3_mux#
fuses for selection of pt3 as a clock source.